[methods] Move lints from strings.rs to methods module (#14253)
