+++ clippy_lints/src/declared_lints.rs
@@ -480,14 +480,19 @@ pub static LINTS: &[&crate::LintInfo] = &[
     crate::methods::SLICED_STRING_AS_BYTES_INFO,
     crate::methods::STABLE_SORT_PRIMITIVE_INFO,
     crate::methods::STRING_EXTEND_CHARS_INFO,
+    crate::methods::STRING_FROM_UTF8_AS_BYTES_INFO,
+    crate::methods::STRING_LIT_AS_BYTES_INFO,
     crate::methods::STRING_LIT_CHARS_ANY_INFO,
     crate::methods::STR_SPLIT_AT_NEWLINE_INFO,
+    crate::methods::STRING_TO_STRING_INFO,
+    crate::methods::STR_TO_STRING_INFO,
     crate::methods::SUSPICIOUS_COMMAND_ARG_SPACE_INFO,
     crate::methods::SUSPICIOUS_MAP_INFO,
     crate::methods::SUSPICIOUS_OPEN_OPTIONS_INFO,
     crate::methods::SUSPICIOUS_SPLITN_INFO,
     crate::methods::SUSPICIOUS_TO_OWNED_INFO,
     crate::methods::TYPE_ID_ON_BOX_INFO,
+    crate::methods::TRIM_SPLIT_WHITESPACE_INFO,
     crate::methods::UNBUFFERED_BYTES_INFO,
     crate::methods::UNINIT_ASSUMED_INIT_INFO,
     crate::methods::UNIT_HASH_INFO,
@@ -703,12 +708,7 @@ pub static LINTS: &[&crate::LintInfo] = &[
     crate::string_patterns::SINGLE_CHAR_PATTERN_INFO,
     crate::strings::STRING_ADD_INFO,
     crate::strings::STRING_ADD_ASSIGN_INFO,
-    crate::strings::STRING_FROM_UTF8_AS_BYTES_INFO,
-    crate::strings::STRING_LIT_AS_BYTES_INFO,
     crate::strings::STRING_SLICE_INFO,
-    crate::strings::STRING_TO_STRING_INFO,
-    crate::strings::STR_TO_STRING_INFO,
-    crate::strings::TRIM_SPLIT_WHITESPACE_INFO,
     crate::strlen_on_c_strings::STRLEN_ON_C_STRINGS_INFO,
     crate::suspicious_operation_groupings::SUSPICIOUS_OPERATION_GROUPINGS_INFO,
     crate::suspicious_trait_impl::SUSPICIOUS_ARITHMETIC_IMPL_INFO,
